requirements.txt
python-dotenv>=1.0
streamlit>=1.36
langchain>=0.2
langchain-google-genai>=2.0
langchain-community>=0.2
langchain-text-splitters>=0.2
chromadb>=0.5
pandas>=2.0
